clean up module and import structure
