--- utils.py
@@ -1,38 +1,0 @@
-"""Useful tools."""
-from collections.abc import Mapping
-
-
-class FrozenDict(Mapping):
-    """
-    Immutable dictionary-like type. Found at:
-    https://stackoverflow.com/questions/2703599/what-would-a-frozen-dict-be
-    """
-
-    def __init__(self, *args, **kwargs):
-        self._d = dict(*args, **kwargs)
-        self._hash = None
-
-    def __iter__(self):
-        return iter(self._d)
-
-    def __len__(self):
-        return len(self._d)
-
-    def __getitem__(self, key):
-        return self._d[key]
-
-    def __hash__(self):
-        # It would have been simpler and maybe more obvious to
-        # use hash(tuple(sorted(self._d.iteritems()))) from this discussion
-        # so far, but this solution is O(n). I don't know what kind of
-        # n we are going to run into, but sometimes it's hard to resist the
-        # urge to optimize when it will gain improved algorithmic performance.
-        if self._hash is None:
-            hash_ = 0
-            for pair in self.items():
-                hash_ ^= hash(pair)
-            self._hash = hash_
-        return self._hash
-
-    def __repr__(self):
-        return f"FrozenDict{self._d.__repr__()}"
